feat: optimize ORPC client for SSR performance
- Add server-side client pre-initialization in orpc.server.ts
- Remove async imports from ORPC client for zero-latency SSR
- Eliminate router bundle inclusion on client side
- Improve error handling for missing server client
- Update documentation for new architecture
