+++ kit/dapp/src/orpc/orpc-client.ts
@@ -3,12 +3,16 @@
  *
  * This module exports the primary ORPC client used throughout the application.
  * It creates an isomorphic client that works both on the server and client side:
- * - Server-side: Uses direct router client with request headers from TanStack Start
+ * - Server-side: Uses pre-initialized globalThis.$client for optimal SSR performance
  * - Client-side: Uses OpenAPI link with automatic cookie inclusion for authentication
+ *
+ * The server-side client is initialized in @/lib/orpc.server
+ * to ensure zero client bundle impact while providing optimal SSR performance.
  *
  * The client is integrated with TanStack Query for data fetching and caching.
  * @see {@link ./routes/contract} - Type-safe contract definitions
  * @see {@link ./routes/router} - Main router with all endpoints
+ * @see {@link @/lib/orpc.server} - Server-side client initialization
  */
 
 import { bigDecimalSerializer } from "@atk/zod/bigdecimal";
@@ -17,53 +21,35 @@ import { timestampSerializer } from "@atk/zod/timestamp";
 import { createORPCClient } from "@orpc/client";
 import { RPCLink } from "@orpc/client/fetch";
 import type { RouterClient } from "@orpc/server";
-import { createRouterClient } from "@orpc/server";
 import { createTanstackQueryUtils } from "@orpc/tanstack-query";
-import { createLogger } from "@settlemint/sdk-utils/logging";
 import { createIsomorphicFn } from "@tanstack/react-start";
-import { getHeaders } from "@tanstack/react-start/server";
-import type { router as AppRouter } from "./routes/router";
-
-const logger = createLogger();
+import type { router } from "./routes/router";
 
 /**
  * Creates an isomorphic ORPC client that adapts based on the runtime environment.
  *
  * Server-side behavior:
- * - Creates a direct router client for optimal performance
- * - Automatically includes request headers from TanStack Start
- * - Bypasses HTTP layer for direct function calls
+ * - Uses pre-initialized globalThis.$client from @/lib/orpc.server
+ * - Provides zero-latency direct function calls during SSR
+ * - Eliminates async imports and HTTP overhead
  *
  * Client-side behavior:
  * - Creates an OpenAPI client that communicates via HTTP
  * - Automatically includes cookies for session-based authentication
  * - Points to the `/api` endpoint relative to the current origin
+ * - Router code is completely excluded from client bundle
  */
 const getORPCClient = createIsomorphicFn()
-  .server(async () => {
-    const { router } = await import("./routes/router");
-    return createRouterClient(router, {
-      context: () => {
-        try {
-          const headers = getHeaders();
-          return {
-            headers,
-          };
-        } catch (error) {
-          // Handle cases where there's no HTTP event in AsyncLocalStorage
-          // This can happen during hydration or when there's no active request
-          logger.warn(
-            "No HTTPEvent found in AsyncLocalStorage, using empty headers",
-            { error }
-          );
-          return {
-            headers: {} as ReturnType<typeof getHeaders>,
-          };
-        }
-      },
-    });
+  .server(() => {
+    // Use the pre-initialized server client from @/lib/orpc.server
+    if (!globalThis.$client) {
+      throw new Error(
+        "Server ORPC client not initialized. Ensure @/lib/orpc.server is imported before this module."
+      );
+    }
+    return globalThis.$client;
   })
-  .client((): RouterClient<typeof AppRouter> => {
+  .client((): RouterClient<typeof router> => {
     const link = new RPCLink({
       url: `${globalThis.location.origin}/api/rpc`,
       async fetch(url, options) {
@@ -83,6 +69,6 @@ const getORPCClient = createIsomorphicFn()
     return createORPCClient(link);
   });
 
-export const client = await getORPCClient();
+export const client = getORPCClient();
 
 export const orpc = createTanstackQueryUtils(client);
